* set up work for bento test initialization * split:   * accordion   * base-carousel   * brightcove   * facebook
