+++ build-system/tasks/unit.js
@@ -32,12 +32,13 @@ class Runner extends RuntimeTestRunner {
  * @return {Promise<void>}
  */
 async function unit() {
+  const {bento_only: bentoOnly} = argv;
   maybePrintArgvMessages();
-  if (argv.local_changes && !(await getUnitTestsToRun())) {
+  if (argv.local_changes && !(await getUnitTestsToRun({bentoOnly}))) {
     return;
   }
 
-  const config = new RuntimeTestConfig('unit');
+  const config = new RuntimeTestConfig('unit', {bentoOnly});
   const runner = new Runner(config);
 
   await runner.setup();
@@ -51,6 +52,7 @@ module.exports = {
 
 unit.description = 'Run unit tests';
 unit.flags = {
+  'bento_only': 'Run only bento standalone tests',
   'chrome_canary': 'Run tests on Chrome Canary',
   'chrome_flags': 'Use the given flags to launch Chrome',
   'coverage': 'Run tests in code coverage mode',
