+++ test/fixtures/e2e/bento/fit-text.html
@@ -1,0 +1,26 @@
+<!DOCTYPE html>
+<html>
+
+<head>
+  <script type="module" async src="https://cdn.ampproject.org/bento.mjs"></script>
+
+  <script type="module" async src="https://cdn.ampproject.org/v0/bento-fit-text-1.0.mjs"></script>
+  <link rel="stylesheet" type="text/css" href="https://cdn.ampproject.org/v0/bento-fit-text-1.0.css" />
+
+</head>
+
+<body>
+
+  <bento-fit-text style="
+        width: 300px;
+        height: 300px;
+        border: 1px solid black;
+        display: block;
+      ">
+    Lorem <i>ips</i>um dolor sit amet, has nisl nihil convenire et, vim at
+    aeque inermis reprehendunt.
+  </bento-fit-text>
+
+</body>
+
+</html>
